aws-cluster: Add Admin ACL to the config and properly update them.
Before this patch every aws cluster had a security group created for
it with the IP of the local system a hard coded subnet.  Holes that
were punched were never removed, and the only way to add a new hole
was to recreate the cluster.  With this patch, Admin IP is part of the
config and can be changed at runtime.