config: Generate cloud-config.yaml from config.json
In the future in order to get etcd running properly, we'll need to
change the discovery token in the cloud-config for each new cluster.
This means we'll need to be able to generate it programmatically.
This patch takes a first step in this direction by generating
cloud-config.yaml in the config package.