cluster: Collect all state into single "table" abstraction
This patch collects all of the state strewn about cluster and foreman
into a single "table".  This table is responsible for computing a diff
between the current state of the world, and the proposed state of the
world.  By putting all this logic in one place, we simplify the code,
and pave the way for a more advanced policy engine.