db: Add a new central database for managing policy state.
This patch centralizes all global state into a central database
package, "db".  It adds an "engine" module which is responsible for
updating the database so it conforms with the policy.  That done, the
"cluster" and "container" modules can further augment the database
with details about the instantiation of the specified policy.  This
greatly simplifies the code, and will serve as a platform for the next
generation of complex policies.