db: Tweak the db connection interface
This patch tweaks the database connection interface in an attempt to
make it cleaner and safer.  It makes three changes:

1) New() instantiates all tables instead of requiring callers to know
in advance which tables they'll need.

2) Database modification functions require a handle on the
db.Database.  This ensures that they can only be called during a
transaction.

3) Table rows no longer container a pointer to their parent table.
This allows reflect.DeepEqual() to operate on them and generally
simplifies the code.