dsl: Support lazy evaluation for some function arguments
For some functions, it makes sense to evaluate the arguments after,
not before, the function call.  This particularly makes sense for
makeList which will be used to instantiate atoms.  By waiting to
evaluate the atom arguments until they've been expanded by makeList,
we will actually get the requested number of atoms, instead of just
references to the evaluated versions eagerly evaluated earlier.