dsl: makeList should evaluate its first argument eagerly
Before this patch, did not evaluate it's first argument, making it
impossible to use variables as a count.  This patch changes it so that
statements like the following are possible:

(let ((a 3)) (makeList a a))

In order to do this, func() losses it's concept of eager vs lazy
function arguments.  Instead, functions are responsible for evaluating
their arguments when they deem appropriate.