minion: Ship code instead of data to the minions
In previous patches, the global controller compiled the spec, and
shipped the data necessary for minions to operate.  This patch,
instead, ships the code the minions need to generate their
specification.  In future this will make it easier to offload
computation to the minions, and will simplify the protobuf API.