dsl: Use array indexes instead of pointers for container evaluation
Previously, evaluated atoms would carry with them a pointer to the
container they created in the evalCtx.  This worked poorly when
specifications were evaluated multiple times, as evaluations would
update the Containers in the original evalCtx.  This patch resolves
the issue by using an index int he evalCtx slice instead.