minion: Create new etcd synchronized label table
With this patch, the cluster leader is responsible for allocating an
IP address to each label and figuring out the set of containers that
implement it.  It writes this information to etcd so that it may be
distributed to all of the minions in the cluster.  This will become
important in future patches where this information will be necessary
to do label based networking.