foreman: Strive to maintain existing minion roles
Before this patch, the foreman had blindly accepted whatever minion
role assignments it had found in the database.  The problem with this,
is that if di restarts, it's likely that each minion will be assigned
a new role, causing a major disruption in the cluster.  This patch
resolves the issue by adding an initialization phase to the foreman.
During this phase, it queries the current role of each minion, and
attempts to assign it this role in the database.