minion: Unify label and container IP network state in Etcd
This patch moves responsibility for allocating container IP addresses
to Etcd much in the same way they had been for Labels.  This allows
any node in the system to discover the IP address of every container
and every label that they might depend on.

Since this patch adds containers to Etcd, this patch now writes the
labels that each container implements, instead of the containers that
each label had implemented.  The former is much closer to how it's
expressed in the policy, simplifying things significantly.