language: Implement "placement" keyword and labeling (in the backend)
The only type available for the "placement" keyword is "exclusive".
Labeling existed but needed to be propogated to the container itself
such that we could use docker swarm affinities to implement "placement".

Also fix the Label tests to reflect the new fields added for this patch,
also create a test for Placement "exclusive"

There is also an issue with the docker api regarding capital letters in
the label so a error message was inserted to that effect.