db: Remove repetitive, boilerplate code
This patch uses reflection to generalize the id(), tt(), and String()
functions. Before, adding a new table required copying an existing
example, and tweaking method names to match the new row. Particularly
troublesome was adding new fields to the struct definition for a row,
but not updating the String() function.