log: Switch log system to logrus
There are an abundant number of choices for the logging system in go.
The most obvious choice is the 'log' package in the standard library,
which seems like a reasonable system, but is missing a lot of features
we're going to need.  Specifically, the ability to have multiple log
levels that can be turned off and on as needed.

Logrus seems to be popular in this community (it's used by docker),
and has some nice properties including optional machine readable log
output and color.  We're going with it for now.