cluster: Move provider logic to its own package
Clusters and providers are no longer as intertwined as they were before.
As we move towards easy support of multiple providers (potentially
within the same cluster), it makes more sense to move the provider logic
to its own package.
This patch also tweaks the semantics of initializing providers. There is
now a `provider.New` function that returns an empty instance of a given
provider, and a `Start` method that must be implemented to fully initialize
the provider. This way, we can define "static" methods in the `Provider`
interface that don't cause side-effects from creating an empty
`Provider` instance.
Cloud configs are also no longer exposed to the cluster anymore, and are
managed internally by the providers.