provider: Use `di` user for di tasks on the minion
It was confusing that the di user was different between AWS and Vagrant
(`ubuntu` vs. `vagrant`), so this patch explicitly creates a user named
`di`.

This also makes the Vagrant code a little cleaner because we don't need
to explicitly add the Vagrant private key to our authorized keys
anymore. Because we don't modify the default vagrant user, `vagrant ssh`
works with no modification.