store: Fix a panic with unlabeled containers
If a container was unlabeled the `Label` column in the container table
would be nil, confusing the network store code and causing a panic.
Upon inspection, the same could happen if a container was deleted and
the removal from etcd failed for some reason.  This patch fixes the
issue by dealing simply treating this case like the empty list.