ovsdb: Clean up helpers and add new accessor functions
Cleaned up the mutation and condition helper functions as well as adding
a function for parsing libovsdb maps. Added utilities for manipulating
"openflow ports" and "openflow interfaces" which are the ports and
interfaces in OVS. Also added a special error to indicate that something
was not found, along with ways to detect it.