network: Don't assign special IPs to single container labels
When a label has a single container, there's no reason to assign it a
special IP address independent of that container.  In fact, doing so
confuses certain applications, in particular zookeeper.  This patch
makes a special case to alleviate the issue.  It also updates the
zookeeper spec file to take advantage of the new option.