dsl: Ship unevaluated code to the minion
Before this patch, the dsl would ship the code post evaluation to the
minion.  This caused a lot of trouble as we had to ensure that the
code post evaluation was equivalent to the code pre-evaluation even in
the presence of side effects.  This patch solves the problem by
shipping the pre-evaluated code instead.  I order to achieve this, the
patch introduces a pre-processing stage `resolveImports()` that
collapses multiple module files into a single specification.  This
allows us to ignore imports in the minion.