dsl: Allow a list of labels in connect from
This patch tweaks connect so it can accept a list of of labels as the
from argument.  It then pairwise connects every label in this from
list, to every label in the to list.

In order to achieve this some significant changes were necessary for
the unit tests.  Specifically, code comparisons are now whitespace
agnostic.  Also, we no longer evaluate code twice to verify it works
on the minions as the minions now receive unevaluated code.