dsl: Make labels first-class citizens
This allows labels to be returned from functions. To achieve this,
labelable types (machines and containers) were also made first class
citizens.
Furthermore, this scheme is more correct for labeling things within lambda
functions. Before, if an atom were created anywhere but the global
scope, there was potential for collisions. For example, a lambda
function may return an `astAtom` with index 0; however, the atom array
of the outside scope would be different, so index 0 would point to a
different atom! This isn't a problem now because labels point directly
to atoms, without this indexing scheme inbetween.