network: Public Internet connections
With this patch, containers no longer connect to the internet by
default.  The policy can now specify connections to/from the public
internet by using the `connect` keyword and the special label
"public".  For now, a connection to/from the public web allows both
tcp and udp traffic on the specified port. Furthermore, all ICMP
traffic is allowed between the public web and containers that
implement labels connected *to* the public web.

The current implementation does not allow the policy to specify port
ranges for connections that involve the public internet. This is
because OpenFlow doesn't easily allow port ranges.

We need exclusive placement for containers that allow ingoing
connections from the public web on the same port.