cluster: Sleep in the event of a provider error.
If a provider returns an error, such as specifically an API request
limit exceeded error,  the cluster will often immediately try again
which can in some cases exacerbate the problem.  This patch deals with
it by forcing the provider to sleep for one minute in the event of an
error.  This is somewhat of a stop gap safety mechanism, future
patches will address the underlying issue.