minion: Respect labels when syncing containers
Before, we only used the image, command, and environment to correspond
database containers with running docker containers. This patch adds
labels to the criteria for equality.

This was a problem when the scheduler booted containers with the same
image, command, and environment, but different labels:
```
(label "client" (docker "iperf" "-s"))
(label "server" (docker "iperf" "-s"))
(connect 5001 "client" "server")
```
Half the time, the engine would swap the client and server, and enforce
the connection on the wrong container.