network: Fix typo that causes etcd thrashing
In minion/network/store.go we had a check to ensure we only write to
ETCD if the IP address of a label changed.  This check had a typo
causing it to fail.  As a result, every time the network code woke up,
it wrote all IPs to etcd causing serious performance problems.