scheduler: Move check for running supervisor containers
The supervisor ensures that the `etcd`, `ovs`, `swarm`, etc containers
are running with the proper parameters by attempting to start the
containers, regardless of whether or not they're already running.
Before, the docker `create` function checked whether the container was
already running. Now, that logic is moved into the supervisor, which
allows us to remove the expensive `getID` call from the docker `create`
function.