inspect: Invariant checking and placement awareness
An invariant is a condition about a stitch that must hold for
the stitch to be valid. The inspect tool can now check a set
of invariants specified in an external file.
The inspect tool, when generating visualizations, now is aware
of placement directives in the stitch and will display
containers that will be scheduled together in groups.