join: Add linear time HashJoin
Joins are more-or-less ubiquitous in the code and some of them,
particularly those operating over ACLs and containers, operate over a
great deal of data.  Before this patch, all Joins uses an O(n^2)
algorithm which scaled badly.  This patch adds a new HashJoin()
function that operates in O(n) time for those use cases that have
simple score functions.