spark: Allow passing in a job to the Spark specification
An important use case for running Spark on Quilt will be the ability
to run Spark jobs reproducibly on many different infrastructures.  To
this end, this patch adds an example stitch, sparkPI, that boots a
spark cluster, waits for it to come up healthy, and then uses it to
calculate PI.  To achieve this, the master and worker scripts were
replaced with a more sophisticated python script and the examples were
added to the spark Dockerfile.