amazon: Create security groups in all regions
Before this patch, the Amazon EC2 provider would only try to create
security groups in regions that had a session.  Sessions weren't
created until a call to Boot().  This, of course, presented a problem
because booting a machine requires a security group to be already
created.  This cased a rather cryptic error message that resolved
itself given long enough.  This patch fixes the problem by having the
SetACL() function range over all regions, not just those that have an
already created session.