docs: Simplify example.spec
This patch makes various tweaks to example.spec in an effort to make
it as simple as possible.  Specifically it relies more heavily on
using Quilt defaults, and removes Amazon specific configuration.
After these changes the example spec can be modified to operate on a
different cloud provider by simply changing the `Provider` variable.

The patch also moves example.spec to the docs directory, and updates
GettingStarted.md to link to its new location.