quilt: Add a helper utility for stopping infrastructures
Before this patch, if one wanted to destroy an infrastructure, she
would have to modify the specification to remove all of the
infrastructure VMs.  This patch adds a helper utility that installs an
empty Stitch in a specified namespace, making it easy to stop all of
the VMs in that namespace.

Suggested-By: Vivian Fang <vivian98.fang@gmail.com>
Suggested-By: Steven H. Wang <wang.steven.h@gmail.com>