stitch: Introduce Container IDs
Before this patch, in the stitch internals, there was no way to refer
to a container except through the labels that point to it.  This
causes a number of rather deep issues.  Most notably, all operations
on containers, `connect, `place`, etc, must happen on labels instead
of directly on their underlying containers.  By giving containers IDs
within stitch, we'll be able to revamp placement in the language, and
in future patches, make label members individually addressable within
the dataplane.