cluster: Rewrite cluster.go and foreman.go
The cluster package was over designed to support multiple clusters
running at a time, which is something that we don't need at the
moment. Changes have been made to remove this functionality that
simplifies some portions of the code, and after that, further and
more general changes were made in order to clean up the code.
Additionally, the foreman used to run as a goroutine that was
spawned by the cluster. This separation led to both increased
complexity and the inability of the foreman to know when it was
able to start doing its job. This has also been changed, and now
foreman acts as a library that cluster can call at the appropriate
times.