stitch: Handle public port exclusivity internally
We require that all labels listening on a given public port be placed
exclusively with each other (and themselves).  Before this patch, this
functionality was handled in the minion engine by adding `PortRule`s.
This patch implements the functionality entirely within stitch using
the existing label exclusivity functionality.  This simplifies a lot
of the code significantly.