stitch: Simplify placement representation
This patch simplifies the representation of placements exposed by
stitch in the QueryPlacements function.  First, it gets rid of a
placement Rule, instead embedding its fields in the Placement struct
directly.  Second, it changes the OtherLabel field from a slice of
strings to a string.  If a placement rule applies to multiple labels,
then it will be expanded into multiple Placement structs.  Doing all
this greatly simplifies future patches.