db: Simplify db table alert system
Instead of listening through a channel in a goroutine, Transact now
uses a mutex lock to ensure safe concurrent access. Also, instead of
tracking the table seq, we simply flag tables as needing to be
alerted or not, which allows us to ditch row.equal and just use
reflect.DeepEqual for better correctness with minimal performance
hits.